+++ src/app/modules/workspace/workspace.service.ts
@@ -6,7 +6,7 @@ import { createLoggerWithLabel } from '../../config/logger'
 import { getWorkspaceModel } from '../../models/workspace.server.model'
 import { transformMongoError } from '../../utils/handle-mongo-error'
 import { DatabaseError, DatabaseValidationError } from '../core/core.errors'
-import * as FormService from '../form/form.service'
+import * as AdminFormService from '../form/admin-form/admin-form.service'
 
 import {
   ForbiddenWorkspaceError,
@@ -93,22 +93,24 @@ export const deleteWorkspace = ({
   workspaceId: string
   userId: string
   shouldDeleteForms: boolean
-}): ResultAsync<boolean, DatabaseError> => {
+}): ResultAsync<void, DatabaseError> => {
   return ResultAsync.fromPromise(
     WorkspaceModel.startSession().then((session: ClientSession) =>
-      session.withTransaction(() =>
-        deleteWorkspaceTransaction({
-          workspaceId,
-          userId,
-          shouldDeleteForms,
-          session,
-        }).then(() => session.endSession()),
-      ),
+      session
+        .withTransaction(() =>
+          deleteWorkspaceTransaction({
+            workspaceId,
+            userId,
+            shouldDeleteForms,
+            session,
+          }),
+        )
+        .then(() => session.endSession()),
     ),
-
     (error) => {
       logger.error({
-        message: 'Database error when deleting workspace',
+        message:
+          'Database error when deleting workspace, rolling back transaction',
         meta: {
           action: 'deleteWorkspace',
           workspaceId,
@@ -117,7 +119,7 @@ export const deleteWorkspace = ({
         },
         error,
       })
-      return error as ReturnType<typeof transformMongoError>
+      return transformMongoError(error)
     },
   )
 }
@@ -132,52 +134,25 @@ const deleteWorkspaceTransaction = async ({
   userId: string
   shouldDeleteForms: boolean
   session: ClientSession
-}): Promise<boolean> => {
-  const logMeta = {
-    action: 'deleteWorkspaceTransaction',
-    workspaceId,
-    userId,
-    shouldDeleteForms,
-  }
-  let isDeleteSuccessful = false
+}): Promise<void> => {
   const workspaceToDelete = await WorkspaceModel.findOne({
     _id: workspaceId,
     admin: userId,
   })
 
-  try {
-    isDeleteSuccessful = await WorkspaceModel.deleteWorkspace({
-      workspaceId,
-      admin: userId,
-      session,
-    })
-  } catch (error) {
-    logger.error({
-      message: 'Error while deleting workspace, rolling back transaction',
-      meta: logMeta,
-      error,
-    })
-    throw transformMongoError(error)
-  }
+  await WorkspaceModel.deleteWorkspace({
+    workspaceId,
+    admin: userId,
+    session,
+  })
 
   if (shouldDeleteForms && workspaceToDelete?.formIds) {
-    try {
-      await FormService.archiveForms({
-        formIds: workspaceToDelete?.formIds,
-        userId,
-        session,
-      })
-    } catch (error) {
-      logger.error({
-        message: 'Error while archiving forms, rolling back transaction',
-        meta: logMeta,
-        error,
-      })
-      throw transformMongoError(error)
-    }
+    await AdminFormService.archiveForms({
+      formIds: workspaceToDelete?.formIds,
+      userId,
+      session,
+    })
   }
-
-  return isDeleteSuccessful
 }
 
 export const getForms = (
